Remove hb-ot-shape-complex-machine-index.hh file
Inline code into USE header.
